Implement first-order rule cost calculation
This is an attempt to take into account the full knowledge
of the ruleset into the calculation of the cost of a rule.
It's not perfect since it works only on the 'first order',
that is to say, it only adds the costs of directly 'implied'
children, not of it's children, and it also adds the costs
of nonspecific children. I don't yet know how to improve
that without causing an infinite loop.